[Feat]: Creating custom hook
- created custom hook for useFetch.
- that will mainly seperate the fetch function from the app.js to useFetch,
- Like function we can use that custom hook multiple times.
