+++ place_picker/src/App.jsx
@@ -7,6 +7,7 @@ import logoImg from './assets/logo.png';
 import AvailablePlaces from './components/AvailablePlaces.jsx';
 import { fetchUserPlaces, updateUserPlaces } from './http.js';
 import Error from './components/Error.jsx';
+import useFetch from './hooks/useFetch.js';
 
 function App() {
   const selectedPlace = useRef();
@@ -19,24 +20,8 @@ function App() {
 
   const [modalIsOpen, setModalIsOpen] = useState(false);
 
-  // fetch user places 
-  useEffect(() => {
-    async function fetchPlaces() {
-      setIsFetching(true);
-      try {
-        const places= await fetchUserPlaces();
-        setUserPlaces(places);
-      } catch (error) {
-        setError({
-          message: error.message || 'Failed to fetch user places.'
-        })
-      }
-
-      setIsFetching(false);
-    }
-
-    fetchPlaces();
-  }, []);
+  // custom hook for useFetch 
+  useFetch();
 
   function handleStartRemovePlace(place) {
     setModalIsOpen(true);
